+++ packages/reusable-blocks/src/store/selectors.js
@@ -1,10 +1,9 @@
 /**
- * Returns true if reusable block is in the editing state.
+ * Returns false because the deprecated reusable blocks package no longer tracks
+ * editing state.
  *
- * @param {Object} state    Global application state.
- * @param {number} clientId the clientID of the block.
  * @return {boolean} Whether the reusable block is in the editing state.
  */
-export function __experimentalIsEditingReusableBlock( state, clientId ) {
-	return state.isEditingReusableBlock[ clientId ];
+export function __experimentalIsEditingReusableBlock() {
+	return false;
 }