+++ packages/reusable-blocks/src/store/test/selectors.js
@@ -4,24 +4,7 @@
 import { __experimentalIsEditingReusableBlock } from '../selectors';
 
 describe( '__experimentalIsEditingReusableBlock', () => {
-	it( 'gets the value for clientId', () => {
-		expect(
-			__experimentalIsEditingReusableBlock(
-				{ isEditingReusableBlock: { 1: true } },
-				1
-			)
-		).toBe( true );
-		expect(
-			__experimentalIsEditingReusableBlock(
-				{ isEditingReusableBlock: { 2: false } },
-				2
-			)
-		).toBe( false );
-		expect(
-			__experimentalIsEditingReusableBlock(
-				{ isEditingReusableBlock: { 2: false } },
-				3
-			)
-		).toBe( undefined );
+	it( 'returns false', () => {
+		expect( __experimentalIsEditingReusableBlock( {}, 1 ) ).toBe( false );
 	} );
 } );